Move FoundProperty from LogProperties to LogProperty
FoundProperty (and its StringProperty/ListProperty/MapProperty variants)
belongs with the Result machinery that consumes it, not with the raw
LogProperties source interface - moved verbatim, including the
StringProperty-local maybeRedact helper. LogProperties.FoundProperty
references become LogProperty.FoundProperty throughout.

stringPropertyOrNull/listPropertyOrNull/mapPropertyOrNull stay on
LogProperties: investigated removing them since forKey(...) covers the
common case, but they're still load-bearing for composite property sources
(ListLogProperties overrides all three to preserve which *specific* child
LogProperties a value actually came from - see FileOutputPropertiesTest's
golden error strings, where the "key: 'x' from Y" portion names one exact
source while the separate "Tried:" line lists every source searched).
Falling back to the default valueOrNull()-based implementation for a
composite would collapse that distinction and regress those messages.
A single unified replacement doesn't work either: MutableLogProperties'
mapOrNull() override can find dotted sub-keys even when valueOrNull(key)
itself is null, so the three accessors can't share one has-a-value check.
