+++ core/src/main/java/io/jstach/rainbowgum/LogProperty.java
@@ -4,6 +4,7 @@ import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -13,7 +14,6 @@ import java.util.function.Supplier;
 
 import org.eclipse.jdt.annotation.Nullable;
 
-import io.jstach.rainbowgum.LogProperties.FoundProperty;
 import io.jstach.rainbowgum.LogProperty.PropertyFunction;
 import io.jstach.rainbowgum.LogProperty.PropertyValue;
 import io.jstach.rainbowgum.LogProperty.Result;
@@ -342,6 +342,104 @@ public interface LogProperty {
 			throw new IllegalArgumentException("Property key should not start or end with '" + LogProperties.SEP + "'");
 		}
 		LogProperties.validateKeyParameters(key, Set.of());
+	}
+
+	/**
+	 * Found property retrieved from {@link LogProperties}. This is a bridge and meta data
+	 * needed for the {@link LogProperty} fluent like monads. It includes the original
+	 * value before conversions.
+	 *
+	 * @apiNote This sealed class is purposely not generic parameterized but you are
+	 * allowed to pattern match as the subclasses represent the builtin types of
+	 * properties that are supported.
+	 */
+	public sealed interface FoundProperty {
+
+		/**
+		 * The originating <em>exact</em> properties that the value was found on.
+		 * @return properties.
+		 */
+		LogProperties properties();
+
+		/**
+		 * The key that was used to find this property.
+		 * @return key also known as property name.
+		 */
+		String key();
+
+		/**
+		 * A string representation of the value that this property has usually for error
+		 * descriptions.
+		 * @return description of value.
+		 */
+		String valueDescription();
+
+		/**
+		 * A found <strong>string</strong> property result which includes the
+		 * <strong>exact</strong> properties where a value was found.
+		 *
+		 * @param properties the <strong>exact</strong> properties where the value was
+		 * found.
+		 * @param key property key.
+		 * @param value property string value.
+		 */
+		public record StringProperty(LogProperties properties, String key, String value) implements FoundProperty {
+			@Override
+			public String valueDescription() {
+				return maybeRedact(value);
+			}
+
+			private static final Set<String> REDACTED_KEYS = Set.of("password", "apikey", "secret", "token");
+
+			private static final String REDACTED_VALUE = "<REDACTED>";
+
+			private static final String maybeRedact(String input) {
+				String lower = input.toLowerCase(Locale.ROOT);
+				if (REDACTED_KEYS.contains(lower)) {
+					return REDACTED_VALUE;
+				}
+				for (var k : REDACTED_KEYS) {
+					if (input.contains(k)) {
+						return REDACTED_VALUE;
+					}
+				}
+				return input;
+			}
+		}
+
+		/**
+		 * A found <strong>list</strong> property result which includes the
+		 * <strong>exact</strong> properties where a value was found.
+		 *
+		 * @param properties the <strong>exact</strong> properties where the value was
+		 * found.
+		 * @param key property key.
+		 * @param value property string value.
+		 */
+		public record ListProperty(LogProperties properties, String key, List<String> value) implements FoundProperty {
+			@Override
+			public String valueDescription() {
+				return StringProperty.maybeRedact("" + value);
+			}
+		}
+
+		/**
+		 * A found <strong>map</strong> property result which includes the
+		 * <strong>exact</strong> properties where a value was found.
+		 *
+		 * @param properties the <strong>exact</strong> properties where the value was
+		 * found.
+		 * @param key property key.
+		 * @param value property string value.
+		 */
+		public record MapProperty(LogProperties properties, String key,
+				Map<String, String> value) implements FoundProperty {
+			@Override
+			public String valueDescription() {
+				return StringProperty.maybeRedact("" + value);
+			}
+		}
+
 	}
 
 	/**
